src/requirements.txt
psycopg2
google-cloud-secret-manager
google-cloud-bigquery
sqlalchemy
pandas
mysql-connector-python
db-dtypes
